\name{glass}
\alias{glass}
\non_function{}
\title{Glass Identification Database}
\description{
The \code{glass} data frame has 214 rows and 10 columns, representing
glass fragments.
}
\format{
This data frame contains the following columns:
}
\arguments{
\item{RI}{
refractive index
}
\item{Na}{
weight percent in corresponding oxide
}
\item{Mg}{
weight percent in corresponding oxide
}
\item{Al}{
weight percent in corresponding oxide
}
\item{Si}{
weight percent in corresponding oxide
}
\item{K}{
weight percent in corresponding oxide
}
\item{Ca}{
weight percent in corresponding oxide
}
\item{Ba}{
weight percent in corresponding oxide
}
\item{Fe}{
weight percent in corresponding oxide
}
\item{Type}{
Type of glass:
1 building\_windows\_float\_processed,
2 building\_windows\_non\_float\_processed,
3 vehicle\_windows\_float\_processed,
4 vehicle\_windows\_non\_float\_processed (none in this database),
5 containers,
6 tableware,
7 headlamps
}
}
\source{
    P. M. Murphy and D. W. Aha (1999),
    UCI Repository of Machine Learning Databases,
    \url{ftp://ics.uci.edu/pub/machine-learning-databases}
}
