\name{mda-internal}
\alias{contr.fda}
\alias{fix.theta}
\alias{kmeans.start}
\alias{laplacian}
\alias{llmult}
\alias{lvq.start}
\alias{make.dumpdata.mda}
\alias{mean.penalty}
\alias{mda.fit}
\alias{mda.means}
\alias{polybasis}
\alias{pplot}
\alias{pplot4}
\alias{shrink}
\alias{shrink.mda}
\alias{transform.penalty}
\title{Internal mda Functions}
\description{Internal mda functions.}
\details{
  These are not to be called by the user.
}
\keyword{utilities}
