\name{predict.mars}
\alias{predict.mars}
\title{
    make predictions from a `mars' object
}
\usage{
predict.mars(object, x)
}
\arguments{
    \item{object}{a \code{mars} object}
    \item{x}{values at which predictions are to be made.}
}
\value{
    the fitted values
}
\seealso{
    \code{\link{mars}},
    \code{\link{predict}},
    \code{\link{model.matrix.mars}}
}
\examples{
## see help for mars()
}
% Converted by Sd2Rd version 0.3-3.
